requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
crashtest==0.4.1
cryptography==39.0.0
distlib==0.3.6
Django==4.1.5
dulwich==0.20.50
filelock==3.9.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.0.0
msgpack==1.0.4
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
ptyprocess==0.7.0
pycparser==2.21
pyrsistent==0.19.3
python-dotenv==0.21.1
python-telegram-bot==20.0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.1.20
urllib3==1.26.14
virtualenv==20.17.1
webencodings==0.5.1
zipp==3.11.0
